Extract `factorial` to shared utility to remove duplication
- Moved `factorial` function from `Math/Numerical_Methods/Functions/Factorial/factorial.py` to `Math/utils/math_utils.py` and consolidated implementations.
- Updated imports in combinatorics and trigonometric files to point to the shared `factorial` implementation.
- Fixed `sys.path` hacks in `combination.py` and `trinomial_theorem_general_term.py`.
- Updated test expectations in `test_cosine_rule.py` to match the exact string raised by the new utility function.
