+++ Math/Discrete_Math/Combinatorics/combination.py
@@ -1,9 +1,6 @@
 # Combination formula: for nCr where n and r are non-negative integers
 
-import sys
-import os
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../../..')))
-from Math.Numerical_Methods.Functions.Factorial.factorial import factorial
+from Math.utils.math_utils import factorial
 
 def nCr(n: int, r: int) -> int:
     """
