+++ Math/Geometry/Trigonometry/Formulas/cosine_rule.py
@@ -1,19 +1,7 @@
 # Cosine rule (Law of Cosines) for finding sides and angles of triangles
 from typing import Union
 
-
-def factorial(n: int) -> int:
-    """Calculate factorial of n."""
-    if n < 0:
-        raise ValueError("Factorial not defined for negative numbers.")
-    if n > 1000:
-        raise ValueError(
-            "Factorial calculation limit exceeded (maximum allowed is 1000)."
-        )
-    result = 1
-    for i in range(1, n + 1):
-        result *= i
-    return result
+from Math.utils.math_utils import factorial
 
 
 def cosine_taylor(radians: Union[int, float]) -> float:
